+++ python_src/requirements.txt
@@ -2,3 +2,5 @@ torch
 torchvision
 tqdm
 black[d]
+polars
+requests
